Add config file using a toml format
