+++ .gitignore
@@ -4,3 +4,4 @@
 Cargo.lock
 state.txt
 scratch.chs
+*config.toml
