+++ Cargo.toml
@@ -11,4 +11,5 @@ pest = "2.7.3"
 pest_derive = "2.7.3"
 reqwest = { version = "0.11", features = ["blocking", "json"] }
 serde_json = "1.0"
-serde = "1.0.193"
+serde = {  version = "1.0.193", features = ["derive"] }
+toml = "0.8.12"
